# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/telegram-settings": {
            "get": {
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "description": "Get the Telegram bot settings. The bot token is never returned; has_bot_token reports whether one is stored.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Get Telegram settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "description": "Save the Telegram bot settings. An empty bot_token keeps the stored one.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Save Telegram settings",
                "parameters": [
                    {
                        "description": "Telegram settings",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TelegramSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/telegram-settings/test": {
            "post": {
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "description": "Send a test message to the configured default chat",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Test Telegram bot",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.TelegramSettings": {
            "type": "object",
            "properties": {
                "bot_token": {
                    "type": "string"
                },
                "default_chat_id": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "notify_next_race": {
                    "type": "boolean"
                },
                "notify_results": {
                    "type": "boolean"
                },
                "reminder_days": {
                    "type": "string"
                },
                "reminder_hour": {
                    "type": "integer"
                },
                "subscribers": {
                    "type": "integer"
                }
            }
        },
...
    }
}